Add files via upload
This file stores the Inventory information. 
Everything from item description, weight, quantity, etc. 
There is more inventory to add. 
Ultimately, it should work in conjunction with the Inventory Program. 
Main goal of program is to be able to help with maintaining Inventory accuracy at 99.9% accurate. 

THIS IS ONLY A PROTOTYPE.